requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.2
tortoise-orm
aiomysql
pydantic~=2.11.4
app~=0.0.1